requirements.txt
# Django Betting Platform - Requirements
# Python 3.8+

# Core Django
Django==4.2.7
djangorestframework==3.14.0

# Database
mysqlclient==2.2.0
pymysql==1.1.0

# Email
django-sendgrid-v5==1.2.2
sendgrid==6.11.0

# Authentication & Security
django-allauth==0.57.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Development
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Data Migration
pandas==2.1.3
